+++ NEWS.md
@@ -25,6 +25,7 @@ New features
   - ``;`` can now be used to indicate the insertion of the minimal move number `` . `` which
     has been interpreted to mean the previous move number with an extra ``.`` tacked to the end.
   - Can now have identical ``MoveNumbers`` but will emit a warning if identical ``MoveNumbers`` are found.
+  - Can now directly set piece attributes for suit, rank, angle, and configuration (#35)
 
 Bug fixes and minor improvements
 --------------------------------